♻️ refactor: scan/result 컴포넌트 분리 및 스타일 코로케이션
