+++ src/components/scan/result/SaveBottomSheet.tsx
@@ -1,27 +1,19 @@
-import { memo, useCallback, useEffect, useMemo, useRef, useState } from 'react';
+import { useCallback, useEffect, useMemo, useRef, useState } from 'react';
 import {
   View,
-  Text,
-  TextInput,
-  TouchableOpacity,
   Modal,
   Pressable,
   Animated,
   StyleSheet,
   Keyboard,
-  ScrollView,
   Platform,
   useWindowDimensions,
-  ActivityIndicator,
   Alert,
 } from 'react-native';
-import { Ionicons } from '@expo/vector-icons';
 import { useSafeAreaInsets } from 'react-native-safe-area-context';
 import { useTranslation } from 'react-i18next';
 import i18n from '../../../i18n';
-import { PrimaryButton, SecondaryButton } from '../../common/Button';
-import styles from '../../../styles/scan/saveBottomSheet';
-import colors from '../../../constants/colors';
+import styles from './styles';
 import {
   getCalendarPreview,
   postCalendarEvents,
@@ -30,6 +22,9 @@ import {
 } from '../../../api/calendar';
 import type { CalendarPreviewItem } from '../../../api/calendar';
 import { getNewsletterTranslation } from '../../../api/newsletter';
+import type { EventState } from './EventPreviewCard';
+import ConfirmStep from './ConfirmStep';
+import SuccessStep from './SuccessStep';
 
 interface Props {
   visible: boolean;
@@ -41,47 +36,7 @@ interface Props {
   newsletterTitle?: string;
 }
 
-type EventState = {
-  year: string;
-  month: string;
-  day: string;
-  isEditing: boolean;
-};
-
-interface DateInputFieldsProps {
-  year: string;
-  month: string;
-  day: string;
-  onYearChange: (v: string) => void;
-  onMonthChange: (v: string) => void;
-  onDayChange: (v: string) => void;
-}
-
-interface EventCardProps {
-  item: CalendarPreviewItem;
-  es: EventState;
-  childName: string;
-  onUpdate: (id: string, patch: Partial<EventState>) => void;
-  onDateConfirm: (id: string) => void;
-  getDisplayDate: (y: string, m: string, d: string) => string;
-}
-
 const SHEET_HEIGHT = 560;
-const STYLE_FULL_WIDTH = { width: '100%' } as const;
-const STYLE_FLEX_1 = { flex: 1 } as const;
-const STYLE_FLEX_2 = { flex: 2 } as const;
-
-const localStyles = StyleSheet.create({
-  scrollArea: { width: '100%', flex: 1 },
-  scrollContent: { gap: 20, paddingBottom: 4 },
-  loadingIndicator: { marginVertical: 16 },
-  errorText: {
-    fontSize: 13,
-    color: colors.text.secondary,
-    textAlign: 'center',
-    paddingVertical: 8,
-  },
-});
 
 const formatCorrectedDate = (y: string, m: string, d: string): string | null => {
   const yn = Number(y);
@@ -100,143 +55,6 @@ const mapRegisterError = (e: unknown): string => {
   }
   return i18n.t('scan.result.saveBottomSheet.error.registerDefault');
 };
-
-const DateInputFields = memo(
-  ({ year, month, day, onYearChange, onMonthChange, onDayChange }: DateInputFieldsProps) => {
-    const { t } = useTranslation();
-    return (
-      <View style={styles.dateInputRow}>
-        <View style={styles.dateInputWrap}>
-          <TextInput
-            style={styles.dateInput}
-            value={year}
-            onChangeText={onYearChange}
-            keyboardType="number-pad"
-            maxLength={4}
-            accessibilityLabel={t('scan.result.saveBottomSheet.yearLabel')}
-          />
-          <Text style={styles.dateUnit}>{t('scan.result.saveBottomSheet.year')}</Text>
-        </View>
-        <View style={styles.dateInputWrap}>
-          <TextInput
-            style={styles.dateInput}
-            value={month}
-            onChangeText={onMonthChange}
-            keyboardType="number-pad"
-            maxLength={2}
-            accessibilityLabel={t('scan.result.saveBottomSheet.monthLabel')}
-          />
-          <Text style={styles.dateUnit}>{t('scan.result.saveBottomSheet.month')}</Text>
-        </View>
-        <View style={styles.dateInputWrap}>
-          <TextInput
-            style={styles.dateInput}
-            value={day}
-            onChangeText={onDayChange}
-            keyboardType="number-pad"
-            maxLength={2}
-            accessibilityLabel={t('scan.result.saveBottomSheet.dayLabel')}
-          />
-          <Text style={styles.dateUnit}>{t('scan.result.saveBottomSheet.day')}</Text>
-        </View>
-      </View>
-    );
-  }
-);
-
-const EventCard = memo(
-  ({ item, es, childName, onUpdate, onDateConfirm, getDisplayDate }: EventCardProps) => {
-    const { t } = useTranslation();
-    const displayDate = getDisplayDate(es.year, es.month, es.day);
-
-    const handleEditToggle = useCallback(
-      () => onUpdate(item.tempEventId, { isEditing: !es.isEditing }),
-      [onUpdate, item.tempEventId, es.isEditing]
-    );
-    const handleConfirm = useCallback(
-      () => onDateConfirm(item.tempEventId),
-      [onDateConfirm, item.tempEventId]
-    );
-    const handleYearChange = useCallback(
-      (v: string) => onUpdate(item.tempEventId, { year: v }),
-      [onUpdate, item.tempEventId]
-    );
-    const handleMonthChange = useCallback(
-      (v: string) => onUpdate(item.tempEventId, { month: v }),
-      [onUpdate, item.tempEventId]
-    );
-    const handleDayChange = useCallback(
-      (v: string) => onUpdate(item.tempEventId, { day: v }),
-      [onUpdate, item.tempEventId]
-    );
-
-    return (
-      <View style={styles.eventCard}>
-        <View style={styles.eventHeader}>
-          <View style={styles.eventDot} />
-          <Text style={styles.eventTitle}>
-            {item.title} · {childName}
-          </Text>
-        </View>
-        {item.isDateExtracted ? (
-          <View style={styles.eventDateRow}>
-            <Text style={styles.eventDate}>{displayDate}</Text>
-            <TouchableOpacity
-              style={styles.editBadge}
-              onPress={handleEditToggle}
-              accessibilityRole="button"
-              accessibilityLabel={t('scan.result.saveBottomSheet.accessibilityEdit')}
-            >
-              <Text style={styles.editBadgeText}>
-                {es.isEditing
-                  ? t('scan.result.saveBottomSheet.isEditing')
-                  : t('scan.result.saveBottomSheet.edit')}
-              </Text>
-            </TouchableOpacity>
-          </View>
-        ) : (
-          <>
-            <View style={styles.divider} />
-            <DateInputFields
-              year={es.year}
-              month={es.month}
-              day={es.day}
-              onYearChange={handleYearChange}
-              onMonthChange={handleMonthChange}
-              onDayChange={handleDayChange}
-            />
-          </>
-        )}
-        {item.isDateExtracted && es.isEditing && (
-          <View style={styles.dateEditorCard}>
-            <Text style={styles.dateEditorLabel}>
-              {t('scan.result.saveBottomSheet.dateEditor')}
-            </Text>
-            <DateInputFields
-              year={es.year}
-              month={es.month}
-              day={es.day}
-              onYearChange={handleYearChange}
-              onMonthChange={handleMonthChange}
-              onDayChange={handleDayChange}
-            />
-            <TouchableOpacity
-              style={styles.dateConfirmBtn}
-              onPress={handleConfirm}
-              activeOpacity={0.8}
-              accessibilityRole="button"
-              accessibilityLabel={t('scan.result.saveBottomSheet.accessibilityDateConfirm')}
-            >
-              <Text style={styles.dateConfirmBtnText}>
-                {t('scan.result.saveBottomSheet.dateConfirm')}
-              </Text>
-            </TouchableOpacity>
-          </View>
-        )}
-      </View>
-    );
-  }
-);
 
 const SaveBottomSheet = ({
   visible,
@@ -509,128 +327,30 @@ const SaveBottomSheet = ({
               <View style={styles.handle} />
 
               {step === 'success' ? (
-                <>
-                  <ScrollView
-                    style={localStyles.scrollArea}
-                    contentContainerStyle={localStyles.scrollContent}
-                    showsVerticalScrollIndicator={false}
-                    keyboardShouldPersistTaps="handled"
-                  >
-                    <View style={styles.iconWrap}>
-                      <Ionicons name="calendar-outline" size={28} color={colors.primary[400]} />
-                    </View>
-                    <View style={styles.textBlock}>
-                      <Text style={styles.title}>
-                        {t('scan.result.saveBottomSheet.successTitle')}
-                      </Text>
-                      <Text style={styles.subtitle}>
-                        {t('scan.result.saveBottomSheet.successSubtitle')}
-                      </Text>
-                    </View>
-                    {previews.map((p) => {
-                      const es = eventStates[p.tempEventId];
-                      return (
-                        <View key={p.tempEventId} style={styles.eventCard}>
-                          <View style={styles.eventHeader}>
-                            <View style={styles.eventDot} />
-                            <View style={styles.successEventInfo}>
-                              <Text style={styles.eventTitle}>
-                                {p.title} · {childName}
-                              </Text>
-                              {es && (
-                                <Text style={styles.eventDate}>
-                                  {getDisplayDate(es.year, es.month, es.day)}
-                                </Text>
-                              )}
-                            </View>
-                          </View>
-                        </View>
-                      );
-                    })}
-                  </ScrollView>
-                  <PrimaryButton
-                    label={t('scan.result.saveBottomSheet.viewCalendar')}
-                    onPress={onConfirm}
-                    style={STYLE_FULL_WIDTH}
-                  />
-                  <SecondaryButton
-                    label={t('scan.result.saveBottomSheet.close')}
-                    onPress={onDismiss}
-                    style={STYLE_FULL_WIDTH}
-                  />
-                </>
+                <SuccessStep
+                  previews={previews}
+                  eventStates={eventStates}
+                  childName={childName}
+                  getDisplayDate={getDisplayDate}
+                  onConfirm={onConfirm}
+                  onDismiss={onDismiss}
+                />
               ) : (
-                <>
-                  <ScrollView
-                    style={localStyles.scrollArea}
-                    contentContainerStyle={localStyles.scrollContent}
-                    showsVerticalScrollIndicator={false}
-                    keyboardShouldPersistTaps="handled"
-                  >
-                    <View style={styles.iconWrap}>
-                      <Ionicons name="calendar-outline" size={28} color={colors.primary[400]} />
-                    </View>
-                    <View style={styles.textBlock}>
-                      <Text style={styles.title}>{t('scan.result.saveBottomSheet.title')}</Text>
-                      <Text style={styles.subtitle}>
-                        {hasAnyMissingDate
-                          ? t('scan.result.saveBottomSheet.subtitleManualDate')
-                          : t('scan.result.saveBottomSheet.subtitleAutoDate')}
-                      </Text>
-                    </View>
-
-                    {hasAnyMissingDate && (
-                      <View style={styles.warningCard}>
-                        <Ionicons name="warning" size={16} color={colors.text.primary} />
-                        <Text style={styles.warningText}>
-                          {t('scan.result.saveBottomSheet.dateNotFound')}
-                        </Text>
-                      </View>
-                    )}
-
-                    {previewLoading && (
-                      <ActivityIndicator
-                        size="small"
-                        color={colors.primary[400]}
-                        style={localStyles.loadingIndicator}
-                      />
-                    )}
-                    {!previewLoading && previewError && (
-                      <Text style={localStyles.errorText}>{previewError}</Text>
-                    )}
-                    {!previewLoading &&
-                      !previewError &&
-                      previews.map((p) => {
-                        const es = eventStates[p.tempEventId];
-                        if (!es) return null;
-                        return (
-                          <EventCard
-                            key={p.tempEventId}
-                            item={p}
-                            es={es}
-                            childName={childName}
-                            onUpdate={updateEventState}
-                            onDateConfirm={handleDateConfirmFor}
-                            getDisplayDate={getDisplayDate}
-                          />
-                        );
-                      })}
-                  </ScrollView>
-
-                  <View style={styles.buttons}>
-                    <SecondaryButton
-                      label={t('scan.result.saveBottomSheet.no')}
-                      onPress={onClose}
-                      style={STYLE_FLEX_1}
-                    />
-                    <PrimaryButton
-                      label={registering ? '...' : t('scan.result.saveBottomSheet.register')}
-                      onPress={handleRegisterConfirm}
-                      disabled={registering || !canRegister}
-                      style={STYLE_FLEX_2}
-                    />
-                  </View>
-                </>
+                <ConfirmStep
+                  hasAnyMissingDate={hasAnyMissingDate}
+                  previewLoading={previewLoading}
+                  previewError={previewError}
+                  previews={previews}
+                  eventStates={eventStates}
+                  childName={childName}
+                  registering={registering}
+                  canRegister={canRegister}
+                  onUpdate={updateEventState}
+                  onDateConfirm={handleDateConfirmFor}
+                  getDisplayDate={getDisplayDate}
+                  onClose={onClose}
+                  onRegister={handleRegisterConfirm}
+                />
               )}
             </View>
           </Animated.View>
